+++ libraries/exoplayer_ima/src/main/java/androidx/media3/exoplayer/ima/ImaServerSideAdInsertionUriBuilder.java
@@ -25,6 +25,7 @@ import androidx.annotation.Nullable;
 import androidx.media3.common.C;
 import androidx.media3.common.C.ContentType;
 import androidx.media3.common.util.UnstableApi;
+import com.google.ads.interactivemedia.v3.api.CustomUiOptions;
 import com.google.ads.interactivemedia.v3.api.ImaSdkFactory;
 import com.google.ads.interactivemedia.v3.api.StreamRequest;
 import com.google.ads.interactivemedia.v3.api.StreamRequest.StreamFormat;
@@ -58,6 +59,10 @@ public final class ImaServerSideAdInsertionUriBuilder {
   private static final String STREAM_ACTIVITY_MONITOR_ID = "streamActivityMonitorId";
   private static final String FORMAT = "format";
   private static final String LOAD_VIDEO_TIMEOUT_MS = "loadVideoTimeoutMs";
+  private static final String CUSTOM_UI_OPTIONS_SKIPPABLE_SUPPORT =
+      "customUiOptionsSkippableSupport";
+  private static final String CUSTOM_UI_OPTIONS_ABOUT_THIS_AD_SUPPORT =
+      "customUiOptionsAboutThisAdSupport";
 
   @Nullable private String adsId;
   @Nullable private String assetKey;
@@ -70,6 +75,7 @@ public final class ImaServerSideAdInsertionUriBuilder {
   @Nullable private String authToken;
   @Nullable private String streamActivityMonitorId;
   private ImmutableMap<String, String> adTagParameters;
+  @Nullable private CustomUiOptions customUiOptions;
   public @ContentType int format;
   private int loadVideoTimeoutMs;
 
@@ -218,6 +224,19 @@ public final class ImaServerSideAdInsertionUriBuilder {
   }
 
   /**
+   * Sets the custom UI options for the stream request.
+   *
+   * @param customUiOptions Custom UI options for the stream request.
+   * @return This instance, for convenience.
+   */
+  @CanIgnoreReturnValue
+  public ImaServerSideAdInsertionUriBuilder setCustomUiOptions(
+      @Nullable CustomUiOptions customUiOptions) {
+    this.customUiOptions = customUiOptions;
+    return this;
+  }
+
+  /**
    * Sets the optional stream manifest's suffix, which will be appended to the stream manifest's
    * URL. The provided string must be URL-encoded and must not include a leading question mark.
    *
@@ -321,6 +340,15 @@ public final class ImaServerSideAdInsertionUriBuilder {
           AD_TAG_PARAMETERS, adTagParametersUriBuilder.build().toString());
     }
     dataUriBuilder.appendQueryParameter(FORMAT, String.valueOf(format));
+    if (customUiOptions != null) {
+      CustomUiOptions customUiOptions = this.customUiOptions;
+      dataUriBuilder.appendQueryParameter(
+          CUSTOM_UI_OPTIONS_SKIPPABLE_SUPPORT,
+          String.valueOf(customUiOptions.getSkippableSupport()));
+      dataUriBuilder.appendQueryParameter(
+          CUSTOM_UI_OPTIONS_ABOUT_THIS_AD_SUPPORT,
+          String.valueOf(customUiOptions.getAboutThisAdSupport()));
+    }
     return dataUriBuilder.build();
   }
 
@@ -400,6 +428,24 @@ public final class ImaServerSideAdInsertionUriBuilder {
     @Nullable String streamActivityMonitorId = uri.getQueryParameter(STREAM_ACTIVITY_MONITOR_ID);
     if (streamActivityMonitorId != null) {
       streamRequest.setStreamActivityMonitorId(streamActivityMonitorId);
+    }
+
+    @Nullable
+    String customUiOptionsSkippableSupport =
+        uri.getQueryParameter(CUSTOM_UI_OPTIONS_SKIPPABLE_SUPPORT);
+    @Nullable
+    String customUiOptionsAboutThisAdSupport =
+        uri.getQueryParameter(CUSTOM_UI_OPTIONS_ABOUT_THIS_AD_SUPPORT);
+    if (customUiOptionsSkippableSupport != null || customUiOptionsAboutThisAdSupport != null) {
+      CustomUiOptions customUiOptions = ImaSdkFactory.createCustomUiOptions();
+      if (customUiOptionsSkippableSupport != null) {
+        customUiOptions.setSkippableSupport(Boolean.parseBoolean(customUiOptionsSkippableSupport));
+      }
+      if (customUiOptionsAboutThisAdSupport != null) {
+        customUiOptions.setAboutThisAdSupport(
+            Boolean.parseBoolean(customUiOptionsAboutThisAdSupport));
+      }
+      streamRequest.setCustomUiOptions(customUiOptions);
     }
     return streamRequest;
   }